+++ mathf.h
@@ -13,8 +13,8 @@
 #include <string.h>
 #include "struct.h"
 
-float Polynomial(poly_s polyf, float in);
-void Rectangular(float* values, int size, float stepsize, float* integ1, float* integ2);
-float Trapezoidal(float* values, int size, float stepsize);
+double Polynomial(poly_s polyf, double in);
+void Rectangular(double* values, int size, double stepsize, double* integ1, double* integ2);
+float Trapezoidal(double* values, int size, double stepsize);
 
 #endif
